split test-participant run_server() into dds submodule
